+++ src/GloriousConsoleAdventure/Menu/ActionMenu.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using GloriousConsoleAdventure.Models.Hero;
 
-namespace GloriousConsoleAdventure
+namespace GloriousConsoleAdventure.Menu
 {
     public class ActionMenu
     {
